Add multithreading support for concurrent time limit and remove ability retest
Added multithreading support to have the test end immediately after the time limit expires rather than previously where the check for pastTimeLimit would not occur until AFTER the user gave the input to the current question they were working on.

This unfortunately lead to the removal of the ability to retest since it is not simple to end a running thread. Will look into readding this feature by utilizing condition variables to signal to the question asking thread that it needs to stop.